Assignment2/requirements.txt
dnspython==2.8.0
psycopg2-binary==2.9.10
pymongo==4.15.3
python-dotenv==1.1.1
